improves log
